Split OpenAI loop tests
